bede/requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.1
APScheduler==3.10.4
PyYAML==6.0.2
